turn livereload into a standalone middleware
